Rearranged the functions and file directory.
All DB queries and models will be under models package.

All functions to process/extract/save the links will be under helpers.go file.

Added test cases and mocks for database queries functions.
